Experiment lib: Disable analysis for service experiments
Resolves a bug where analyze.py failed because it did not recognize
one of the flags in the conf files.  Additionally, the CSV file output
by a Spark service run seems to have some differences (in particular,
in the WORKER_POOL line) from the CSV file output by a normal
simulator run, which analyze.py can't cope with.  It might be useful
in the future to ensure the format of the CSV file is consistent and
reenable analysis for service runs.