Merge workload spec branch into main (#10)
* Update tpch-spark to include changes from pass-deadlines branch

* Read workload spec json in launch_tpch_queries.py

Changes the interface of launch_tpch_queries.py to accept a JSON file
that contains the specification of the workload to launch.  This
avoids complications with simulator/service parity by making the
workload fully deterministic.

* specify deadline when creating task graph

* Create script to generate a workload spec

* modify loader to read from spec file

* remove comment about metadata field

* Spark service: accept deadlines from TPC-H query info

* Update ray search script to use workload specs

* bug fixes

* add release time to deadline

* corrected setup for config search 3/5

* fix dumb bug in ray search script

* more dumb search script fixes

also only apply the dsched slo penalty calculation in the metric if
dsched's slo < 80%

* Fix bug in service when deadline is set

The deadline needs to be converted to an EventTime; otherwise the call
to get_next_task_graph fails and the task graph is never created.

* Spark service experiment setup for 3/06

* Experiment setup for 03/08

E/M/H mixture; extra datapoints for DSched and EDF

* make tpch_plot work again

* update ray search

* add graphene config

* tpch plot graphene + tetrisched baselines

---------

Co-authored-by: Rohan Bafna <130247393+rohanbafna@users.noreply.github.com>